replacing size_t with unsigned long, so that the names are mangled the same way on x86 and x86_64 + update the def files

+++ src/lib/WP1PictureGroup.cpp
@@ -58,7 +58,7 @@ void WP1PictureGroup::_readContents(WPXInputStream *input, WPXEncryption *encryp
 	input->seek(-2, WPX_SEEK_CUR);
 	for (int i = 0; i < 512; i++)
 		m_binaryData.append((unsigned char)0);
-	for (size_t j = 0; j < dataSize && !input->atEOS(); j++ );
+	for (unsigned long j = 0; j < dataSize && !input->atEOS(); j++ );
 		m_binaryData.append(readU8(input, encryption));
 #if DUMP_PICTURE
 	std::ostringstream filename;
